+++ src/main/php/ClassProxy.php
@@ -22,15 +22,4 @@ interface ClassProxy extends Proxy
      * @throws  \InvalidArgumentException  in case any of the mapped methods does not exist or is not applicable
      */
     public function returns(array $callMap): self;
-
-    /**
-     * alias for returns()
-     *
-     * @api
-     * @deprecated  since 3.2.0, use returns() instead, will likely be removed with 4.0.0
-     * @param   array  $callMap
-     * @return  $this
-     * @throws  \InvalidArgumentException  in case any of the mapped methods does not exist or is not applicable
-     */
-    public function mapCalls(array $callMap): self;
 }